requirements.txt
Appium-Python-Client==4.0.0
pytest==8.2.1
pytest-html==4.1.1
